bug fix: getCell returned "undefined" when it was expected to return -2 or -4
because the methods "getLastRenderedRow" and "getLastRenderedColumn" were incorrectly using rendered index in a comparison where source index was expected

in this commit I am also fixing "getLastVisibleRow" and "getLastVisibleColumn", but I am not adding any tests for it. I am sorry, these methods are not already tested and I don't have more time to work on this. I think that tests for these methods should be added